Create Day7.py
For my first exercise in this course, I wrote a code to create a basic calculator that can perform some fundamental arithmetic operations between two given numbers (they are not input by the user) in Python. The output of the code is displayed in a format that is easy to read and understand.