Add files via upload
In this code, I created a nested tuple which contains tuples having the name and Twitter username of a few of my friends. The code asks the user to input the name of the person whose ID they are searching for and returns it or informs the user that the person is not in the tuple.