Add files via upload
In this code, I created a dictionary containing key-value pairs of all the countries that have won the FIFA WC and their number of trophies. The code asks the user to input which country they choose and then displays the number of titles won by that country and the key-value pairs in the form of a list containing tuples of each key-value pair.